test(observability): split envelope roundtrip coverage
