+++ tests/unit/observability/envelope-roundtrip.test.ts
@@ -494,16 +494,4 @@ describe('envelope roundtrip — AC1 positive', () => {
 		expect(first.eventId).not.toBe(second.eventId);
 		expect(second.writerSequence).toBeGreaterThan(first.writerSequence);
 	});
-
-	test('cross-process fixture (no shared in-memory state between two independent createObservation calls) round-trips', () => {
-		resetObservabilityForTesting();
-		const procA = createObservation('heartbeat', { sessionId: 'proc-a' });
-		resetObservabilityForTesting();
-		const procB = createObservation('heartbeat', { sessionId: 'proc-b' });
-		expect(ObservabilityEventSchema.safeParse(procA).success).toBe(true);
-		expect(ObservabilityEventSchema.safeParse(procB).success).toBe(true);
-		expect(validateEventRelationships(procA).ok).toBe(true);
-		expect(validateEventRelationships(procB).ok).toBe(true);
-		expect(procA.trace.traceId).not.toBe(procB.trace.traceId);
-	});
 });
